+++ src/add_event_listener.ts
@@ -1,4 +1,5 @@
 import { removeEventListener } from './remove_event_listener';
+import { Maybe, maybeOf } from '@fluss/core';
 import type { EventMapOf, EventListenerOrEventListenerObject } from './types';
 
 /**
@@ -15,18 +16,27 @@ export function addEventListener<
   E extends EventTarget,
   T extends keyof EventMapOf<E>
 >(
-  element: E,
+  element: E | Maybe<E>,
   type: T,
   listener: EventListenerOrEventListenerObject<E, T>,
   options: {
     add?: boolean | AddEventListenerOptions;
     remove?: boolean | EventListenerOptions;
   } = {}
-): () => void {
-  // @ts-ignore
+): Maybe<() => void> {
   // keyof returns string | number | symbol type, which is not
   // exists in EventMapOf<E> type. TypeScript possibly cannot narrow type here.
   // Also TypeScript's EventListenerOrEventListenerObject cannot narrow event type :(
-  element.addEventListener(type, listener, options.add);
-  return () => removeEventListener(element, type, listener, options.remove);
+  return maybeOf(element)
+    .map((target) => {
+      target.addEventListener(
+        type as string,
+        listener as EventListener | EventListenerObject,
+        options.add
+      );
+      return target;
+    })
+    .map((target) => () =>
+      removeEventListener<E, T>(target as E, type, listener, options.remove)
+    );
 }