+++ src/clone_node.ts
@@ -1,3 +1,8 @@
-export function cloneNode(node: Node, deep: boolean = false): Node {
-  return node.cloneNode(deep);
+import { Maybe, maybeOf } from '@fluss/core';
+
+export function cloneNode(
+  node: Node | Maybe<Node>,
+  deep: boolean = false
+): Maybe<Node> {
+  return maybeOf(node).map((node) => node.cloneNode(deep));
 }