+++ src/closest.ts
@@ -1,8 +1,10 @@
-import { maybeOf, Maybe } from "@fluss/core";
+import { Maybe, maybeOf } from '@fluss/core';
 
 export function closest<T extends Element>(
   selector: string,
-  child: Element
+  child: Element | Maybe<Element>
 ): Maybe<T> {
-  return maybeOf(child.closest<T>(selector));
+  return maybeOf(child).map((childElement) =>
+    childElement.closest<T>(selector)
+  );
 }