+++ src/create_text_node.ts
@@ -1,5 +1,0 @@
-import { Wrapper, wrap } from "@fluss/core";
-
-export function createTextNode(data: string): Wrapper<Text> {
-  return wrap(document.createTextNode(data));
-}