+++ src/has_attribute.ts
@@ -1,8 +1,11 @@
-import { AttributeNamesOf, GlobalAttributeNames } from './types';
+import { Maybe, maybeOf } from '@fluss/core';
+import type { AttributeNamesOf, GlobalAttributeNames } from './types';
 
 export function hasAttribute<E extends Element>(
-  element: E,
+  element: E | Maybe<E>,
   name: AttributeNamesOf<E> | GlobalAttributeNames
 ): boolean {
-  return element.hasAttribute(name);
+  return maybeOf(element)
+    .map((el) => el.hasAttribute(name))
+    .extract();
 }