+++ src/prepend_nodes.ts
@@ -1,6 +1,8 @@
+import { Maybe, maybeOf } from '@fluss/core';
+
 export function prependNodes(
-  parent: ParentNode,
+  parent: ParentNode | Maybe<ParentNode>,
   ...children: ReadonlyArray<string | Node>
 ): void {
-  parent.prepend(...children);
+  maybeOf(parent).map((parentElement) => parentElement.prepend(...children));
 }