+++ src/query_selector_all.ts
@@ -1,8 +1,14 @@
+import { Maybe, maybeOf } from '@fluss/core';
+
 export function querySelectorAll<T extends Element>(
   selector: string,
-  parent: ParentNode = document
+  parent: ParentNode | Maybe<ParentNode> = document
 ): ReadonlyArray<T> {
   const array: Array<T> = [];
-  parent.querySelectorAll<T>(selector).forEach((item) => array.push(item));
+  maybeOf(parent).map((parentElement) =>
+    parentElement
+      .querySelectorAll<T>(selector)
+      .forEach((item) => array.push(item))
+  );
   return Object.freeze(array);
 }