+++ src/remove_attribute.ts
@@ -1,11 +1,12 @@
 import { hasAttribute } from './has_attribute';
-import { AttributeNamesOf, GlobalAttributeNames } from './types';
+import { Maybe, maybeOf } from '@fluss/core';
+import type { AttributeNamesOf, GlobalAttributeNames } from './types';
 
 export function removeAttribute<E extends Element>(
-  element: E,
+  element: E | Maybe<E>,
   name: AttributeNamesOf<E> | GlobalAttributeNames
 ): void {
   if (hasAttribute(element, name)) {
-    element.removeAttribute(name);
+    maybeOf(element).map((el) => el.removeAttribute(name));
   }
 }