+++ src/remove_node.ts
@@ -1,3 +1,5 @@
-export function removeNode(node: ChildNode): void {
-  node.remove();
+import { Maybe, maybeOf } from '@fluss/core';
+
+export function removeNode(node: ChildNode | Maybe<ChildNode>): void {
+  maybeOf(node).map((n) => n.remove());
 }