+++ test/create_text_node.test.ts
@@ -1,8 +1,0 @@
-import { isWrapper } from '@fluss/core';
-import { createTextNode } from '../src';
-
-describe('createTextNode', () => { 
-  test('createTextNode creates test node wrapped by Wrapper', () => {
-    expect(isWrapper(createTextNode('Hello'))).toBe(true);
-  })
-});